[FIX] Changed Leader behavior for listening to everyone, changed int to size_t for better comparison, same for size_t to int

+++ lib/utils/entry.hh
@@ -16,6 +16,6 @@ public:
 
     json serialize_message() const;
 
-    int term_;
+    size_t term_;
     std::string command_;
 };
